Helper2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase>=2.18.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx>=0.26.0
websockets==12.0
python-docx==1.1.0
requests>=2.31.0
# PaddleOCR-VL dependencies
paddlepaddle-gpu==3.2.1
paddleocr[doc-parser]>=2.9.0
Pillow>=10.0.0
pdf2image>=1.16.0
pypdf>=3.0.0
